Renomeando arquivos e melhorando os commits
